requirements.txt
torch>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0
numpy>=1.24.0
matplotlib>=3.7.0
pillow>=10.0.0
tqdm>=4.65.0
scikit-learn>=1.3.0
sentencepiece>=0.1.99
protobuf>=3.20.0

